requirements.txt
torch>=2.1.0
diffusers>=0.26.0
transformers>=4.40.0
Pillow>=10.0.0
torchvision>=0.23.0
scipy>=1.11.0
tk>=0.1.0
